Fix fonsTextIterNext when string ends with invalid UTF-8 sequence
If fonsTextIterNext was called with a string that ended in an invalid
UTF-8 sequence then it would erroneously report a success status to the
calling function.

In my case this cause repeated updates to the font texture atlas which
caused major rendering slowdowns.

This fixes that by examining the UTF-8 decoder status before returning
from the function with a success status code.