Basic support for .ttc fonts
stb_truetype supports .ttc fonts, but you have to pass in an offset to say which font in the collection you want. Before, we would always pass in 0, which isn't a valid offset, causing .ttc fonts to fail to load. Now, we pass in the offset of the first font in the collection.

stbtt_GetFontOffsetForIndex is defined to return 0 if the font is a normal .ttf, so in that case the behaviour is unchanged.